// data/tasks/smithing/enum.ts
export enum SmeltingCategoryTaskId {
  SMELT_BRONZE_BAR = "SMELT_BRONZE_BAR",
  SMELT_IRON_BAR = "SMELT_IRON_BAR",
  SMELT_STEEL_BAR = "SMELT_STEEL_BAR",
  SMELT_SILVER_BAR = "SMELT_SILVER_BAR",
  SMELT_GOLD_BAR = "SMELT_GOLD_BAR",
  SMELT_MITHRIL_BAR = "SMELT_MITHRIL_BAR",
  SMELT_PLATINUM_BAR = "SMELT_PLATINUM_BAR",
  SMELT_ADAMANT_BAR = "SMELT_ADAMANT_BAR",
}